Extract db logic
